% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempe.R
\name{tempe_arrests}
\alias{tempe_arrests}
\title{Get Tempe police arrest data (last 90 days)}
\usage{
tempe_arrests()
}
\value{
tibble with arrest records
}
\description{
Get Tempe police arrest data (last 90 days)
}
