% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempe.R
\name{tempe_data}
\alias{tempe_data}
\title{Download a Tempe dataset as CSV}
\usage{
tempe_data(key, layer = NULL)
}
\arguments{
\item{key}{Dataset key from tempe_datasets(), or a raw item_id}

\item{layer}{Layer number (only needed if using raw item_id)}
}
\value{
tibble with dataset records
}
\description{
Download a Tempe dataset as CSV
}
