% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempe.R
\name{tempe_search}
\alias{tempe_search}
\title{Search Tempe open data portal}
\usage{
tempe_search(q = "")
}
\arguments{
\item{q}{Search query}
}
\value{
tibble: id, type, title, description, url
}
\description{
Search Tempe open data portal
}
